Recognize `scroll` as an action and observe relative movement
When `scroll` is the action and `observe-scroll` is also a class
added to the observed element, listen to the scroll event when one
or more "scrollers" is intersecting and track its scroll position
relative to where it was originally triggered.

Scroll position is added as a custom property: `--distance-from-trigger`

As an example, given an element with these classes:

* `observe-trigger-10-top-scroll`
* `observe-scroll`

Once the top of the element intersects the top 10% of the viewport,
a `--distance-from-trigger` property will be added to its inline
style with the top of the element's current pixel difference from
that intersection. As it moves up the viewport, the distance will
increase. If it moves down the viewport, the distance will decrease.

Once the element is no longer intersecting (distance of < 0), the
scroll event will be removed.