Recognize new nodes and class name changes on existing nodes
Add a mutation observer for the `add` and `remove` actions to
reset their behavior on an element that was observed, but then
had its observer disconnected once the add or remove was triggered.
(It's meant to be a one time action) This mutation observer watches
for class name changes so that if a previously triggered element
has a change that causes it to be triggerable again, we recognize
that.

Add a mutation observer that watches for new elements and adds new
observers to them.

Add a unique identifier to observed elements so that we can prevent
observing them multiple times.